pyproject.toml
[project]
name = "elastic"
version = "0.1.0"
description = ""
authors = [
    {name = "jannikitin"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "3.11.2"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "pydantic[email] (>=2.10.5,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.37,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "alembic (>=1.14.0,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pre-commit (>=4.0.1,<5.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pytz (>=2024.2,<2025.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
